add Redshift Serverless e2e tests (#40237)
* add Redshift Serverless e2e tests

* update license year in new files

* use a test cleanup func to close local proxy

* adjust test timeouts

* use a longer conn retry interval
